Release v0
NOTE: This is a pre-release! We cannot guarantee the timing in this
version matches BD.

According to our observation in making BD ver subtitles for BD Vol.1,
the Web source from B-Global has the exactly same video as BD, except
that each episode of BD has a trailing 1s blank. The picture of each
frame is identical, which is verified using PSNR. However, the audio of
B-Global has a 106ms delay. This delay is confirmed using spectrum and
Sushi, and is consistent throughout EP04-06.

Currently we haven't got the BDMV of Vol.2. Yusyabu has purchased the BD
but they are still in delivery (-o-!). This version is based on a bold
assumption that EP07-12 of B-Global will also have the identical video
and a consistent audio delay of 106ms. EP01-06 in this version are timed
based on BD; EP07-12 are timed based on B-Global and are adjusted to
offset the delay. The text and effects are suttled in this version and
unlikely to change. We can just pray that the timing matches BD.